access mosiacml_logger in a differnent way that doesn't affect tests
